+++ utilities/ideascale-importer/ideascale_importer/tests/utils_test.py
@@ -1,6 +1,6 @@
 import unittest
 
-import utils
+from ideascale_importer import utils
 
 
 class TestUtils(unittest.TestCase):
